Fix flay test by removing sleep in test case
