+++ tests/regress/expected/test_ctas_no_preload_lib.out
@@ -9,15 +9,6 @@ DISTRIBUTED BY (i);
 \! gpconfig -c shared_preload_libraries -v $(./data/current_binary_name) > /dev/null
 \! gpstop -far > /dev/null
 \c
--- Make sure that the worker has started. 
--- We cannot use wait_for_worker_new_epoch() here because the worker might not
--- have started yet.
-SELECT pg_sleep(1);
- pg_sleep 
-----------
- 
-(1 row)
-
 SET ROLE test;
 -- Init table_size to include the table
 SELECT diskquota.init_table_size_table();
@@ -30,6 +21,12 @@ SELECT diskquota.init_table_size_table();
 \! gpstop -far > /dev/null
 \c
 SET ROLE test;
+SELECT diskquota.wait_for_worker_new_epoch();
+ wait_for_worker_new_epoch 
+---------------------------
+ t
+(1 row)
+
 SELECT tableid::regclass, size, segid FROM diskquota.table_size
 WHERE tableid = 't_without_diskquota'::regclass ORDER BY segid;
        tableid       |  size   | segid
