+++ tests/regress/expected/test_insert_after_drop.out
@@ -16,22 +16,15 @@ HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sur
 INSERT INTO a SELECT generate_series(1,100);
 -- expect insert fail
 INSERT INTO a SELECT generate_series(1,100000);
-SELECT pg_sleep(10);
- pg_sleep 
-----------
- 
+SELECT diskquota.wait_for_worker_new_epoch();
+ wait_for_worker_new_epoch 
+---------------------------
+ t
 (1 row)
 
 INSERT INTO a SELECT generate_series(1,100);
 ERROR:  schema's disk space quota exceeded with name: sdrtbl
 DROP EXTENSION diskquota;
--- sleep 1 second in case of system slow
-SELECT pg_sleep(1);
- pg_sleep 
-----------
- 
-(1 row)
-
 INSERT INTO a SELECT generate_series(1,100);
 DROP TABLE a;
 \c postgres
