+++ tests/regress/expected/test_table_size.out
@@ -3,17 +3,20 @@ create table a(i text) DISTRIBUTED BY (i);
 NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named 'i' as the Greenplum Database data distribution key for this table.
 HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
 insert into a select * from generate_series(1,10000);
-select pg_sleep(2);
- pg_sleep 
-----------
- 
+SELECT diskquota.wait_for_worker_new_epoch();
+ wait_for_worker_new_epoch 
+---------------------------
+ t
 (1 row)
 
-create table buffer(oid oid, relname name, size bigint) DISTRIBUTED BY (i);
-NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named 'oid' as the Greenplum Database data distribution key for this table.
-HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
-insert into buffer select oid, relname, sum(pg_table_size(oid)) from gp_dist_random('pg_class') where oid > 16384 and (relkind='r' or relkind='m') and relname = 'a' group by oid, relname;
-select sum(buffer.size) = diskquota.table_size.size from buffer, diskquota.table_size where buffer.oid = diskquota.table_size.tableid group by diskquota.table_size.size;
+select pg_table_size('a') as table_size;
+ table_size 
+------------
+     819200
+(1 row)
+
+\gset
+select :table_size = diskquota.table_size.size from diskquota.table_size where tableid = 'a'::regclass and segid=-1;
  ?column? 
 ----------
  t
