+++ tests/regress/expected/test_update_db_cache.out
@@ -37,15 +37,10 @@ DISTRIBUTED BY (i);
 CREATE EXTENSION diskquota;
 WARNING:  [diskquota] diskquota is not ready because current database is not empty
 HINT:  please run 'SELECT diskquota.init_table_size_table();' to initialize diskquota
--- Sleep until the worker adds the current db to cache so that it can be found 
--- when DROP EXTENSION.
--- FIXME: We cannot use wait_for_worker_new_epoch() here because 
--- diskquota.state is not clean. Change sleep() to wait() after removing
--- diskquota.state
-SELECT pg_sleep(5);
- pg_sleep 
-----------
- 
+SELECT diskquota.wait('SELECT diskquota.check_cur_db_status(''UNREADY'');');
+ wait 
+------
+ t
 (1 row)
 
 -- Should find nothing since t_no_extension is not recorded.
