+++ tests/regress/expected/test_worker_not_ready.out
@@ -14,10 +14,10 @@ SELECT diskquota.pause();
 
 -- diskquota.wait_for_worker_new_epoch() cannot be used here because 
 -- diskquota.state is not clean.
-SELECT pg_sleep(5);
- pg_sleep 
-----------
- 
+SELECT diskquota.wait('SELECT diskquota.check_cur_db_status(''UNREADY'');');
+ wait 
+------
+ t
 (1 row)
 
 DROP EXTENSION diskquota;
