+++ tests/regress/sql/test_extension.sql
@@ -20,14 +20,11 @@ CREATE DATABASE dbx10 ;
 show max_worker_processes;
 show diskquota.max_workers;
 
-\! sleep 3; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
-
--- FIXME: We need to sleep for a while each time after CREATE EXTENSION and
--- DROP EXTENSION to wait for the bgworker to start or to exit.
+SELECT diskquota.wait_for_worker_new_epoch();
+\!ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx0
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
@@ -46,14 +43,12 @@ SELECT diskquota.init_table_size_table();
 SELECT diskquota.wait_for_worker_new_epoch();
 SELECT diskquota.set_schema_quota('SX', '1MB');
 SELECT diskquota.wait_for_worker_new_epoch();
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 INSERT INTO SX.a values(generate_series(0, 10));
 DROP TABLE SX.a;
 
 \c dbx2
 CREATE EXTENSION diskquota;
 SELECT diskquota.wait_for_worker_new_epoch();
-\! ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
 SELECT diskquota.set_schema_quota('SX', '1MB');
@@ -64,7 +59,6 @@ DROP TABLE SX.a;
 
 \c dbx3
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
@@ -76,7 +70,6 @@ DROP TABLE SX.a;
 
 \c dbx4
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
@@ -88,7 +81,6 @@ DROP TABLE SX.a;
 
 \c dbx5
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
@@ -100,7 +92,6 @@ DROP TABLE SX.a;
 
 \c dbx6
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
@@ -112,7 +103,6 @@ DROP TABLE SX.a;
 
 \c dbx7
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
@@ -124,7 +114,6 @@ DROP TABLE SX.a;
 
 \c dbx8
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 CREATE SCHEMA SX;
 CREATE TABLE SX.a(i int) DISTRIBUTED BY (i);
@@ -136,79 +125,66 @@ DROP TABLE SX.a;
 
 \c dbx9
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 
 \c dbx10
 CREATE EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 SELECT diskquota.wait_for_worker_new_epoch();
 
 \c dbx0
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx1
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx2
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx3
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx4
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx5
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx6
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx7
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx8
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx9
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c dbx10
 SELECT diskquota.pause();
 SELECT diskquota.wait_for_worker_new_epoch();
 DROP EXTENSION diskquota;
-\! sleep 0.5; ps -ef | grep postgres | grep "\[diskquota]" | grep -v grep | wc -l
 
 \c contrib_regression
