+++ tests/regress/sql/test_readiness_logged.sql
@@ -4,14 +4,14 @@ CREATE DATABASE test_readiness_logged;
 CREATE TABLE t (i int) DISTRIBUTED BY (i);
 
 CREATE EXTENSION diskquota;
-SELECT pg_sleep(5); --Wait for the check completes
+SELECT diskquota.wait('SELECT diskquota.check_cur_db_status(''UNREADY'');');
 
 SELECT count(*) FROM gp_toolkit.gp_log_database
 WHERE logmessage = '[diskquota] diskquota is not ready';
 
 \! gpstop -raf > /dev/null
 \c
-SELECT pg_sleep(1); --Wait for the check completes
+SELECT diskquota.wait('SELECT diskquota.check_cur_db_status(''UNREADY'');');
 
 SELECT count(*) FROM gp_toolkit.gp_log_database
 WHERE logmessage = '[diskquota] diskquota is not ready';
