+++ tests/regress/sql/test_table_size.sql
@@ -4,9 +4,8 @@ create table a(i text) DISTRIBUTED BY (i);
 
 insert into a select * from generate_series(1,10000);
 
-select pg_sleep(2);
-create table buffer(oid oid, relname name, size bigint) DISTRIBUTED BY (oid);
+SELECT diskquota.wait_for_worker_new_epoch();
+select pg_table_size('a') as table_size;
+\gset
+select :table_size = diskquota.table_size.size from diskquota.table_size where tableid = 'a'::regclass and segid=-1;
 
-insert into buffer select oid, relname, sum(pg_table_size(oid)) from gp_dist_random('pg_class') where oid > 16384 and (relkind='r' or relkind='m') and relname = 'a' group by oid, relname;
-
-select sum(buffer.size) = diskquota.table_size.size from buffer, diskquota.table_size where buffer.oid = diskquota.table_size.tableid group by diskquota.table_size.size;
